HEAD-1588 manatee tool installation fails when headnode.sh run again HEAD-1592 Would like to tune down config agent logging a bit

+++ scripts/headnode.sh
@@ -777,11 +777,11 @@ echo "==> Copying manatee tools to GZ."
 manatee=$(vmadm list | grep manatee0 | cut -f1 -d' ')
 for file in $(ls /zones/${manatee}/root/opt/smartdc/manatee/bin/sdc/*); do
     tool=$(basename ${file} .js)
-    mv ${file} /opt/smartdc/bin/${tool}
+    [[ -f /opt/smartdc/bin/${tool} ]] || ln -s ${file} /opt/smartdc/bin/${tool}
 done
 for file in $(ls /zones/${manatee}/root/opt/smartdc/manatee/bin/manta/*); do
     tool=$(basename ${file} .js)
-    ln -s ${file} /opt/smartdc/bin/${tool}
+    [[ -f /opt/smartdc/bin/${tool} ]] || ln -s ${file} /opt/smartdc/bin/${tool}
 done
